# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Snow School Scheduler API
  description: スキー・スノーボードスクールのシフト管理システム API
  version: 1.0.0
  contact:
    name: Snow School Scheduler
servers:
  - url: http://localhost:3000
    description: 開発環境
  - url: https://snow-school-scheduler-api.example.com
    description: 本番環境
paths:
...
  /api/instructors:
    get:
      summary: インストラクター一覧取得
      description: インストラクターの一覧を取得（ステータス・部門絞り込み可能）
      tags:
        - インストラクター管理
      parameters:
        - name: status
          in: query
          description: ステータスでフィルタリング
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, RETIRED]
        - name: departmentId
          in: query
          description: 部門IDでフィルタリング（指定された部門の資格を持つインストラクターのみ）
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '200':
          description: インストラクター一覧
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Instructor'
                            - type: object
                              properties:
                                certifications:
                                  type: array
                                  items:
                                    $ref: '#/components/schemas/CertificationSummary'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: インストラクター作成
      description: 新しいインストラクターを作成
      tags:
        - インストラクター管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorInput'
      responses:
        '201':
          description: インストラクター作成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Instructor'
                          - type: object
                            properties:
                              certifications:
                                type: array
                                items:
                                  $ref: '#/components/schemas/CertificationSummary'
                      message:
                        type: string
                        example: 'Instructor operation completed successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/instructors/[id]:
    parameters:
      - $ref: '#/components/parameters/Id'

    get:
      summary: インストラクター詳細取得
      description: 指定されたインストラクターの詳細情報を取得
      tags:
        - インストラクター管理
      responses:
        '200':
          description: インストラクター詳細
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Instructor'
                          - type: object
                            properties:
                              certifications:
                                type: array
                                items:
                                  $ref: '#/components/schemas/CertificationSummary'
                      message:
                        type: string
                        example: 'Instructor operation completed successfully'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: インストラクター更新
      description: 指定されたインストラクターを更新
      tags:
        - インストラクター管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorInput'
      responses:
        '200':
          description: インストラクター更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Instructor'
                          - type: object
                            properties:
                              certifications:
                                type: array
                                items:
                                  $ref: '#/components/schemas/CertificationSummary'
                      message:
                        type: string
                        example: 'Instructor operation completed successfully'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: リソースのID
      schema:
        type: integer
  responses:
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Resource not found'
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Validation failed'
    ServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            success: false
            error: 'Internal server error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DepartmentSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'スキー'
...
    CertificationSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: 'スキー指導員'
        shortName:
          type: string
          example: '指導員'
        organization:
          type: string
          example: 'SAJ'
        department:
          $ref: '#/components/schemas/DepartmentSummary'

    # 基本レスポンス
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
          example: null
    ApiListResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        count:
          type: integer
          example: 10
        error:
          type: string
          nullable: true
          example: null
    ApiErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Invalid invitation token'

    # 部門関連
...
    Instructor:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
          example: '山田'
        firstName:
          type: string
          example: '太郎'
        lastNameKana:
          type: string
          nullable: true
          example: 'ヤマダ'
        firstNameKana:
          type: string
          nullable: true
          example: 'タロウ'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, RETIRED]
          default: ACTIVE
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InstructorInput:
      type: object
      required:
        - lastName
        - firstName
      properties:
        lastName:
          type: string
          example: '山田'
        firstName:
          type: string
          example: '太郎'
        lastNameKana:
          type: string
          nullable: true
          example: 'ヤマダ'
        firstNameKana:
          type: string
          nullable: true
          example: 'タロウ'
        status:
          type: string
          enum: [ACTIVE, INACTIVE, RETIRED]
          default: ACTIVE
        notes:
          type: string
          nullable: true
        certificationIds:
          type: array
          items:
            type: integer
          description: '割り当てる資格IDの配列'
          example: [1, 3, 5]

    # シフト種類関連
...